requirements.txt
# Django and Core Dependencies
Django==4.2.16
python-decouple==3.8
whitenoise==6.6.0
Pillow==10.4.0

# Database
psycopg2-binary==2.9.9

# API & REST Framework
djangorestframework==3.14.0
django-cors-headers==4.3.1

# SSO & JWT Support
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# Forms & UI
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-widget-tweaks==1.5.0

# Expense System Dependencies
django-money==3.5.0
python-magic==0.4.27
openpyxl==3.1.2
xlsxwriter==3.1.9
reportlab==4.0.7
python-dateutil==2.8.2

# Utilities
django-extensions==3.2.3
django-filter==23.4
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
django-debug-toolbar==4.2.0

# Production
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1

# Monitoring & Logging
sentry-sdk==1.38.0

# Security
django-ratelimit==4.1.0
django-axes==6.1.1